add function of speech synthesis
