+++ pyproject.toml
@@ -10,6 +10,7 @@ gradio = "^3.24.1"
 openai = "^0.27.2"
 notebook = "^6.5.3"
 whisper = "^1.1.10"
+gTTS = "^2.3.1"
 
 [tool.poetry.dev-dependencies]
 pysen = {version = "0.10.3", extras = ["lint"]}
